+++ src/takajo.nim
@@ -29,6 +29,7 @@ include takajopkg/listUnusedRules
 include takajopkg/splitCsvTimeline
 include takajopkg/splitJsonTimeline
 include takajopkg/stackCmdlines
+include takajopkg/stackDNS
 include takajopkg/stackLogons
 include takajopkg/stackProcesses
 include takajopkg/stackTasks
@@ -55,6 +56,7 @@ when isMainModule:
     const example_split_csv_timeline = "  split-csv-timeline -t ../hayabusa/timeline.csv [--makeMultiline] -o case-1-csv\p"
     const example_split_json_timeline = "  split-json-timeline -t ../hayabusa/timeline.jsonl -o case-1-json\p"
     const example_stack_cmdlines = "  stack-cmdlines -t ../hayabusa/timeline.jsonl -o cmdlines.csv\p"
+    const example_stack_dns = "  stack-dns -t ../hayabusa/timeline.jsonl -o dns.csv\p"
     const example_stack_logons = "  stack-logons -t ../hayabusa/timeline.jsonl -o logons.csv\p"
     const example_stack_tasks = "  stack-tasks -t ../hayabusa/timeline.jsonl -o tasks.csv\p"
     const example_stack_processes = "  stack-processes -t ../hayabusa/timeline.jsonl -o processes.csv\p"
@@ -74,9 +76,9 @@ when isMainModule:
         examples & example_extract_scriptblocks &
         example_list_domains & example_list_hashes & example_list_ip_addresses & example_list_undetected_evtx & example_list_unused_rules &
         example_split_csv_timeline & example_split_json_timeline &
-        example_stack_cmdlines & example_stack_logons & example_stack_processes &
-        example_stack_tasks &
-        example_sysmon_process_tree &
+        example_stack_cmdlines & example_stack_dns & example_stack_logons & example_stack_processes &
+        example_stack_cmdlines & example_stack_dns & example_stack_logons & example_stack_processes &
+        example_sysmon_process_tree & example_stack_tasks &
         example_timeline_logon & example_timeline_partition_diagnostic & example_timeline_suspicious_processes &
         example_ttp_summary & example_ttp_visualize & example_ttp_visualize_sigma &
         example_vt_domain_lookup & example_vt_hash_lookup & example_vt_ip_lookup
@@ -189,6 +191,15 @@ when isMainModule:
             short = {
                 "ignoreSysmon": 'y',
                 "ignoreSecurity": 'e'
+            }
+        ],
+        [            
+            stackDNS, cmdName = "stack-dns",
+            doc = "stack DNS queries and responses",
+            help = {
+                "output": "save results to a CSV file",
+                "quiet": "do not display the launch banner",
+                "timeline": "Hayabusa JSONL timeline (profile: any besides all-field-info*)",
             }
         ],
         [