+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 **New Features:**
 
-- `ttp-visualize-sigma` command: extracts out TTPs from sigma rules and puts in a JSON file to upload to MITRE ATT&CK Navigator to visualize in a heatmap. (#92) (@fukusuket)
+- `stack-dns` command: stack DNS requests. (#95) (@fukusuket)
 - `stack-processes` command: stack executed processes. (#93) (@fukusuket)
+- `ttp-visualize-sigma` command: extracts out TTPs from sigma rules and puts in a JSON file to upload to MITRE ATT&CK Navigator to visualize in a heatmap. (#92) (@fukusuket)
 
 **Enhancements:**